+++ projects/fusion-ui/components/calendar/v4/public-api.ts
@@ -1,0 +1,3 @@
+export {CalendarV4Component as CalendarComponent} from './calendar-v4.component';
+export * from '@ironsource/fusion-ui/components/calendar/entities';
+export {CalendarService, CalendarComponentConfigurations} from '@ironsource/fusion-ui/components/calendar/common/base';