+++ projects/fusion-ui/components/dropped-list/v4/public-api.ts
@@ -1,0 +1,2 @@
+export * from './dropped-list.component';
+export * from './dropped-list.entities';