+++ projects/fusion-ui/components/layout/v4/components/layout-header/layout-header.component.ts
@@ -1,14 +1,15 @@
 import {ChangeDetectionStrategy, Component, EventEmitter, HostBinding, Input, Output} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {DynamicComponentsModule} from '@ironsource/fusion-ui/components/dynamic-components/v1';
-import {HeaderContent, TeleportWrapperElement} from '../../layout.entities';
+import {HeaderContent, TeleportSkeleton, TeleportWrapperElement} from '../../layout.entities';
 import {IconModule} from '@ironsource/fusion-ui/components/icon/v1';
 import {IconButtonComponent} from '@ironsource/fusion-ui/components/button/v4';
+import {SkeletonComponent} from '@ironsource/fusion-ui/components/skeleton';
 
 @Component({
     selector: 'fusion-layout-header',
     standalone: true,
-    imports: [CommonModule, DynamicComponentsModule, IconModule, IconButtonComponent],
+    imports: [CommonModule, DynamicComponentsModule, IconModule, IconButtonComponent, SkeletonComponent],
     templateUrl: './layout-header.component.html',
     styleUrls: ['./layout-header.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush
@@ -39,15 +40,34 @@ export class LayoutHeaderComponent {
     get hasTopLine(): boolean {
         return this.isMultiline && !!this._headerContent?.topRowContent?.show;
     }
+
+    get teleportTopRowBaseSkeletons(): TeleportSkeleton[] {
+        return this._headerContent?.topRowContent?.skeletons ?? [];
+    }
+
+    get teleportTopRowBaseSkeletonsGap(): string {
+        return this._headerContent?.topRowContent?.skeletonsGap;
+    }
+
     get teleportTopRowElements(): TeleportWrapperElement[] {
         return this._headerContent?.topRowContent?.teleportElements ?? [];
     }
     get hasBottomLine(): boolean {
         return this.isMultiline && !!this._headerContent?.bottomRowContent?.show;
     }
+
     get bottomTopRowElements(): TeleportWrapperElement[] {
         return this._headerContent?.bottomRowContent?.teleportElements ?? [];
     }
+
+    get teleportBottomLineSkeletons(): TeleportSkeleton[] {
+        return this.isMultiline && this._headerContent?.bottomRowContent?.skeletons ? this._headerContent.bottomRowContent.skeletons : [];
+    }
+
+    get teleportBottomLineSkeletonsGap(): string {
+        return this._headerContent?.bottomRowContent?.skeletonsGap;
+    }
+
     get isDrilldown(): boolean {
         return this.isMultiline && !!this._headerContent?.hasBackButton;
     }