+++ projects/fusion-ui/components/table/v3/components/table-row/table-row.component.ts
@@ -15,8 +15,7 @@ import {TableColumn, TableOptions, TableRowExpandEmitter, TableRowMetaData} from
 import {TableService} from '@ironsource/fusion-ui/components/table/common/services';
 import {isNullOrUndefined} from '@ironsource/fusion-ui/utils';
 import {Observable, of} from 'rxjs';
-import {ColumnData} from './column-data';
-import {TableRow} from '@ironsource/fusion-ui/components/table/common/entities';
+import {TableRow, ColumnData} from '@ironsource/fusion-ui/components/table/common/entities';
 import {IconData} from '@ironsource/fusion-ui/components/icon/v1';
 import {TableTestIdModifiers} from '@ironsource/fusion-ui/entities';
 import {TestIdsService} from '@ironsource/fusion-ui/services/test-ids';