+++ projects/fusion-ui/entities/public-api.ts
@@ -1,2 +1,5 @@
 export * from './layout-user';
 export * from './test-ids-modifiers';
+export * from './ab-test-icons';
+export * from './platform-type';
+export * from './application';